Npz file concat (#3)
This PR fixes the concatenation of multiple NPZ files to form a larger spectrogram as LTAS were computed in the legacy OSEkit.

It introduces the split() method for SpectroData objects, which deviates from the behaviour of AudioData.split() as it forces the chunks to be made on frames on which a window of the SFT is centered. This is required to reconstruct a sft from sft parts, as shown in the "It is possible to calculate the SFT of signal parts:" section.

This PR also improves the accuracy of the SpectroFile objects begin/end timestamps by writting the timestamps in the npz file, rather than computing it in a pretty questionable way as it was done before 🥸.