Feature/strptime format list (#258)
# Description

This PR enhances the `OSmOSE.utils.timestamp_utils.strptime_from_text()` function by allowing the user to pass a **list of strings** (rather than a single string) as `datetime_template`.

The **first** template from the list will be used for extracting the timestamp from the input string.

This enhancement will be helpful in the public API (#256) since we might be confronted to both tz-naive and tz-aware timestamps: we'll be able to try to parse tz-aware timestamps first, then tz-naive timestamps if the date strings don't have `%z` fields.

# Examples

```py
    >>> strptime_from_text('D_12_03_21_hour_11:45:10_PM', '%y_%m_%d_hour_%I:%M:%S_%p')
    Timestamp('2012-03-21 23:45:10') # Retrocompatibility with the previous single string input
    >>> strptime_from_text('2016_06_13_14:12.txt', ['%Y_%m_%d_%H:%M%z','%Y_%m_%d_%H:%M'])
    Timestamp('2016-06-13 14:12:00') # First template from string doesn't match (no %z in text), so the second is applied
    >>> strptime_from_text('2016_06_13_14:12+0500.txt', ['%Y_%m_%d_%H:%M%z','%Y_%m_%d_%H:%M'])
    Timestamp('2016-06-13 14:12:00+0500', tz='UTC+05:00') # First template from string matches, so it is applied
```